# docs/openapi-schema.yaml
openapi: 3.0.3
info:
  title: ACRev360 API
  version: 1.0.0
  description: Revenue administration & collection platform for FCT Area Councils.
paths:
...
  /api/v1/bills:
    get:
      operationId: v1_bills_list
      parameters:
      - in: query
        name: consultant_id
        schema:
          type: integer
        description: Filter by the consultant whose user enumerated the bill's payer.
          Narrows within the caller's own scope — it never widens it for a CONSULTANT/REVENUE_OFFICER.
      - in: query
        name: date_from
        schema:
          type: string
          format: date
        description: Issued on/after this date (inclusive)
      - in: query
        name: date_to
        schema:
          type: string
          format: date
        description: Issued on/before this date (inclusive)
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - in: query
        name: payer
        schema:
          type: integer
        description: Filter to one payer's bills
      - in: query
        name: q
        schema:
          type: string
        description: Search by bill reference or payer name
      - in: query
        name: revenue_item_id
        schema:
          type: integer
        description: Bills carrying a line for this revenue item
      - in: query
        name: status
        schema:
          type: string
        description: Filter by bill status
      - in: query
        name: value_max
        schema:
          type: number
        description: Maximum total_amount (inclusive)
      - in: query
        name: value_min
        schema:
          type: number
        description: Minimum total_amount (inclusive)
      - in: query
        name: ward_id
        schema:
          type: integer
        description: Filter by the bill's payer's ward
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBillList'
          description: ''
    post:
      operationId: v1_bills_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueBillRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IssueBillRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IssueBillRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueBillResponse'
          description: ''
...
components:
  schemas:
...
    Bill:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        bill_ref:
          type: string
          readOnly: true
        payer:
          type: integer
          readOnly: true
        payer_ref:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        arrears_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusE25Enum'
          readOnly: true
        due_date:
          type: string
          format: date
          readOnly: true
        superseded_by:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        consultant_name:
          type: string
          readOnly: true
      required:
      - amount_paid
      - arrears_amount
      - balance
      - bill_ref
      - consultant_name
      - created_at
      - due_date
      - full_name
      - id
      - payer
      - payer_ref
      - status
      - superseded_by
      - total_amount
...
    BillLineEntryRequest:
      type: object
      properties:
        revenue_item_id:
          type: integer
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          default: '1.00'
        rate_band_id:
          type: integer
          nullable: true
        rate_tier_id:
          type: integer
          nullable: true
        amount_override:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
      required:
      - revenue_item_id
...
    IssueBillRequest:
      type: object
      properties:
        payer_id:
          type: integer
        due_date:
          type: string
          format: date
        lines:
          type: array
          items:
            $ref: '#/components/schemas/BillLineEntryRequest'
        bill_all_drafts:
          type: boolean
          default: false
        roll_arrears:
          type: boolean
          default: false
        force:
          type: boolean
          writeOnly: true
          default: false
      required:
      - payer_id
    IssueBillResponse:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        bill_ref:
          type: string
          readOnly: true
        payer:
          type: integer
          readOnly: true
        payer_ref:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        arrears_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusE25Enum'
          readOnly: true
        due_date:
          type: string
          format: date
          readOnly: true
        superseded_by:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        consultant_name:
          type: string
          readOnly: true
        superseded_count:
          type: integer
          readOnly: true
      required:
      - amount_paid
      - arrears_amount
      - balance
      - bill_ref
      - consultant_name
      - created_at
      - due_date
      - full_name
      - id
      - payer
      - payer_ref
      - status
      - superseded_by
      - superseded_count
      - total_amount
...
    PaginatedBillList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Bill'
...
    StatusE25Enum:
      enum:
      - ISSUED
      - PART_PAID
      - PAID
      - OVERDUE
      - CANCELLED
      - SUPERSEDED
      type: string
      description: |-
        * `ISSUED` - Issued
        * `PART_PAID` - Part Paid
        * `PAID` - Paid
        * `OVERDUE` - Overdue
        * `CANCELLED` - Cancelled
        * `SUPERSEDED` - Superseded
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT